// projects/codx-ac/src/lib/models/AdvancedPayment.model.ts
export class AdvancedPayment {
    recID: any;
    journalNo: any;
    journalType: any;
    subType: any;
    voucherNo: any;
    voucherDate: any;
    cashBookID: any;
    contractID: any;
    currencyID: any;
    exchangeRate: any;
    multi: any;
    totalAmt: any;
    objectType: any;
    objectID: any;
    payeeID: any;
    payee: any;
    bankAcctID: any;
    bankID: any;
    pmtMethodID: any;
    reasonID: any;
    memo: any;
    memo2: any;
    refDocs: any;
    status: any;
    totalDR: any;
    totalCR: any;
    illegalAmt: any;
    mixedPayment: any;
    isTransfer: any;
    isSettlement: any;
    validated: any;
    attachments: any;
    refType: any;
    refNo: any;
    refID: any;
    postedLayer: any;
    postedDate: any;
    owner: any;
    buid: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
}